+++ addOns/automation/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Support for generating stats
+- Form and JSON-based authentication support
 
 ## [0.10.1] - 2022-01-05
 ### Fixed